131: downgrade tabler version fixing performance issues, sligth refactor of sidebar
